requirements.txt
# Simple cloud deployment requirements
flask==3.1.1
qrcode[pil]==8.2
Pillow==11.3.0
gunicorn==23.0.0
